requirements.txt
amqp==5.0.9
appdirs==1.4.4
argh==0.26.2
asgiref==3.3.2
attrs==19.3.0
Babel==2.9.1
billiard==3.6.4.0
black==19.10b0
celery==5.2.3
certifi==2023.7.22
chardet==3.0.4
click==8.0.3
Django==3.2.23
django-cors-headers==3.4.0
django-environ==0.4.5
django-filter==2.4.0
django-health-check==3.16.3
django-phonenumber-field==4.0.0
django-prometheus==2.1.0
django_redis==4.12.1
django-rest-auth==0.9.5
djangorestframework==3.11.2
djangorestframework-gis==1.0
flake8==3.8.1
future==0.18.3
google-cloud-bigquery==1.27.2
idna==2.10
importlib-metadata==1.7.0
iso6709==0.1.5
isort==4.3.21
kombu==5.2.3
Markdown==3.2.2
mccabe==0.6.1
pathspec==0.8.0
pathtools==0.1.2
phonenumbers==8.12.6
prometheus-client==0.8.0
psycopg2==2.9.9
pycodestyle==2.6.0
pycountry==20.7.3
pyflakes==2.2.0
pytz==2021.3
PyYAML==5.4
rapidpro-python==2.6.1
redis==4.4.4
regex==2020.7.14
requests==2.31.0
responses==0.10.15
sentry-sdk==1.14.0
six==1.15.0
sqlparse==0.4.4
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.18
vine==5.0.0
watchdog==0.10.3
zipp==3.1.0
opencv-python-headless==4.5.2.54
zbar-py==1.0.4
drf-spectacular==0.20.2
beautifulsoup4==4.10.0
Pillow==10.2.0
django-storages==1.12.3
boto3==1.20.24
django-celery-results==2.4.0
google-api-core==1.33.2
